# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for app.",
        "title": "Brasa",
        "contact": {},
        "version": "1.0"
    },
    "host": "3.145.127.73:8080",
    "paths": {
...
        "/api/all-token": {
            "get": {
                "description": "Create a new user in db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "Get all Tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "contract id",
                        "name": "ContractId",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New User Created successfully",
                        "schema": {
                            "$ref": "#/definitions/inter.AllTokensOutput"
                        }
                    },
                    "500": {
                        "description": "Unable to store data in database",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/api/create-token": {
            "post": {
                "description": "Create a new user in db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "Make Tokens",
                "parameters": [
                    {
                        "description": "Data for make a new token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inter.TokenCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New Token Created successfully",
                        "schema": {
                            "$ref": "#/definitions/inter.TokenCreateOutput"
                        }
                    },
                    "500": {
                        "description": "Unable to created a new token",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/api/get-token": {
            "get": {
                "description": "Create a new user in db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "Get specify token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "E-mail from user",
                        "name": "ContractId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Its a uri content in tokens",
                        "name": "UriToken",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New User Created successfully",
                        "schema": {
                            "$ref": "#/definitions/inter.UserOutputController"
                        }
                    },
                    "500": {
                        "description": "Unable to store data in database",
                        "schema": {
                            "$ref": "#/definitions/erros.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/receivables": {
            "get": {
                "description": "Returns all registered receivables",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receivable"
                ],
                "summary": "Get all receivables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Receivable"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/receivable.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new receivable from the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receivable"
                ],
                "summary": "Creates a new receivable",
                "parameters": [
                    {
                        "description": "Receivable data",
                        "name": "receivable",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.Receivable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Receivable"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/receivable.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/receivables/{id}": {
            "get": {
                "description": "Fetches a receivable by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receivable"
                ],
                "summary": "Get a receivable by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Receivable ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Receivable"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/receivable.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/receivables/{id}/early_payment": {
            "get": {
                "description": "Returns options for anticipating the receivable's payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receivable"
                ],
                "summary": "Fetches early payment options for a receivable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Receivable ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Receivable"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/receivable.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/receivables/{id}/payment_date": {
            "put": {
                "description": "Changes the payment date of the given receivable",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Receivable"
                ],
                "summary": "Updates the payment date of a receivable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Receivable ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Payment Date",
                        "name": "payment_date",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Receivable"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/receivable.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "base.BaseReq": {
            "type": "object",
            "properties": {
                "LogData": {
                    "$ref": "#/definitions/base.LogData"
                },
                "ResponseData": {
                    "$ref": "#/definitions/base.ResponseData"
                }
            }
        },
        "base.LogData": {
            "type": "object",
            "properties": {
                "ExecutionTime": {
                    "type": "string"
                },
                "HttpStatusCode": {
                    "type": "integer"
                },
                "IP": {
                    "type": "string"
                },
                "Route": {
                    "type": "string"
                }
            }
        },
        "base.ResponseData": {
            "type": "object",
            "properties": {
                "Response": {}
            }
        },
        "db.Receivable": {
            "type": "object",
            "properties": {
                "agencia": {
                    "type": "string"
                },
                "banco": {
                    "type": "string"
                },
                "chave_pix": {
                    "type": "string"
                },
                "conta_corrente": {
                    "type": "string"
                },
                "data_emissao_titulo": {
                    "type": "string"
                },
                "data_vencimento_titulo": {
                    "type": "string"
                },
                "desconto_antecipacao": {
                    "type": "number"
                },
                "devedor_cnpj": {
                    "type": "string"
                },
                "devedor_nome": {
                    "type": "string"
                },
                "empresa_cnpj": {
                    "type": "string"
                },
                "empresa_email": {
                    "type": "string"
                },
                "empresa_endereco": {
                    "type": "string"
                },
                "empresa_nome": {
                    "type": "string"
                },
                "empresa_telefone": {
                    "type": "string"
                },
                "garantias": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
        "erros.InternalServerError": {
            "type": "object",
            "properties": {
                "BaseReq": {
                    "$ref": "#/definitions/base.BaseReq"
                },
                "InternalServerError": {
                    "$ref": "#/definitions/erros.Message"
                },
                "Response": {}
            }
        },
        "erros.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "inter.AllTokensOutput": {
            "type": "object",
            "properties": {
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/inter.TokenCreateOutput"
                    }
                }
            }
        },
...
        "inter.TokenCreateInput": {
            "type": "object",
            "properties": {
                "assetId": {
                    "description": "Adicionado AssetID",
                    "type": "string"
                },
                "assetValue": {
                    "type": "number"
                },
                "contractId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expireDate": {
                    "type": "string"
                },
                "guarantees": {
                    "description": "Adicionado Guarantees",
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "maxSupply": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nominalValue": {
                    "description": "Adicionado NominalValue",
                    "type": "number"
                },
                "traits": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "inter.TokenCreateOutput": {
            "type": "object",
            "properties": {
                "contractId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentSupply": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "maxSupply": {
                    "type": "integer"
                },
                "metadataUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "traits": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "uriNumber": {
                    "type": "integer"
                }
            }
        },
...
        "inter.UserOutputController": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "receivable.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}